// src/models/enums/filiere.emun.ts
export enum Filiere {
    CyberSecurite = "Cyber Sécurité",
    DataIA = "Data IA",
    DeveloppementWEB = "Développement WEB et Mobile",
    GenieLogiciel = "Génie Logiciel",
    Quitter = "Quitter"
}

export enum FiliereOptions {
    CyberSecurite = 1,
    DataIA = 2,
    DeveloppementWEB = 3,
    GenieLogiciel = 4,
    Quitter = 5
}